requirements.txt
streamlit  
python-dotenv  
openai  
azure-search-documents  
azure-identity  
azure-storage-blob  
pdfplumber  
tiktoken  
azure-ai-formrecognizer  